+++ lib/src/core/exception/phone_number_target_exception.dart
@@ -1,4 +1,4 @@
-// Target mask phone number exception
+/// Target mask phone number exception
 class PhoneNumberTargetException implements Exception {
   const PhoneNumberTargetException();
 }